+++ src/ipyforcegraph/behaviors/selection.py
@@ -8,7 +8,7 @@ from typing import Optional, Tuple, Union
 import ipywidgets as W
 import traitlets as T
 
-from ._base import Behavior
+from ._base import DEFAULT_RANK, Behavior
 
 
 @W.register
@@ -39,6 +39,10 @@ class NodeSelection(Behavior):
         help="the color of selected nodes",
     ).tag(sync=True)
 
+    @T.default("rank")
+    def _default_rank(self) -> Optional[int]:
+        return DEFAULT_RANK.selection
+
 
 @W.register
 class LinkSelection(Behavior):
@@ -67,14 +71,22 @@ class LinkSelection(Behavior):
     ).tag(sync=True)
 
     selected_curvature: float = T.Float(
-        0, help="the curvature of selected links, default: straight links"
+        None,
+        allow_none=True,
+        help="the curvature of selected links, default: ``None`` preserves unselected ``curvature``.",
     ).tag(sync=True)
 
     selected_line_dash: Tuple[float] = W.TypedTuple(
         T.Float(),
-        help="the line-dash of selected links, default: no dashes",
+        default_value=None,
+        allow_none=True,
+        help="the line-dash of selected links, default: ``None`` preserves unselected ``line_dash``",
     ).tag(sync=True)
 
     selected_width: float = T.Float(2, help="the width of selected links").tag(
         sync=True
     )
+
+    @T.default("rank")
+    def _default_rank(self) -> Optional[int]:
+        return DEFAULT_RANK.selection