+++ src/ipyforcegraph/serializers.py
@@ -11,16 +11,16 @@ import pandas as P
 import traitlets as T
 
 HAS_ORJSON = False
-try:
+try:  # pragma: no cover
     import orjson
 
     HAS_ORJSON = True
-except ImportError:
+except ImportError:  # pragma: no cover
     pass
 
 
 def dataframe_to_json(
-    value: P.DataFrame, widget: W.Widget
+    value: Optional[P.DataFrame], widget: W.Widget
 ) -> Optional[Dict[str, memoryview]]:
     """DataFrame JSON serializer."""
     if value is None:
@@ -30,25 +30,25 @@ def dataframe_to_json(
 
     df_data = value.replace({np.nan: None}).to_dict(orient="list")
 
-    if HAS_ORJSON:
+    if HAS_ORJSON:  # pragma: no cover
         df_json = orjson.dumps(df_data)
-    else:
+    else:  # pragma: no cover
         df_json = json.dumps(df_data).encode("utf-8")
 
     return {"buffer": memoryview(N.zstd.compress(df_json))}
 
 
-def dataframe_from_json(value: Dict[str, Any], widget: W.Widget) -> P.DataFrame:
+def dataframe_from_json(value: Any, widget: W.Widget) -> P.DataFrame:
     """DataFrame JSON de-serializer."""
     if value is None:
         return None
 
-    if "buffer" in value:
+    if isinstance(value, dict) and "buffer" in value:
         decompressed = N.zstd.decompress(value["buffer"])
 
-        if HAS_ORJSON:
+        if HAS_ORJSON:  # pragma: no cover
             df_data = orjson.loads(decompressed.decode("utf-8"))
-        else:
+        else:  # pragma: no cover
             df_data = json.loads(decompressed)
     else:
         df_data = value