+++ CHANGELOG.md
@@ -6,7 +6,7 @@
 
 - improves default selection behavior for `LinkShapes.curvature` and
   `LinkShapes.line_dash`
-- adds `line_dash` to all 2d shapes
+- adds `line_dash` to all 2D shapes
 - removes the significance of order in `ForceGraph.behaviors`
   - all node, link, and graph behaviors now have a (sensible default) `.rank` trait
     which determines the order in which they are applied.